Reorganized logging infrastructure to expose be a bit less focused on k8s
