+++ src/logging/k8s_events.rs
@@ -3,36 +3,16 @@
 use std::error::Error;
 
 use crate::client::Client;
-use kube::{
-    core::DynamicObject,
-    runtime::{
-        controller,
-        events::{Event, EventType, Recorder, Reporter},
-        reflector::ObjectRef,
-    },
+use kube::runtime::{
+    controller,
+    events::{Event, EventType, Recorder, Reporter},
 };
 use tracing::Instrument;
 
-/// [`Error`] extensions that help serialize it to helpful [`Event`]s
-///
-/// This should be implemented for reconciler error types.
-pub trait PublishableError: Error {
-    /// `PascalCase`d name for the error category
-    ///
-    /// This can typically be implemented by delegating to [`strum_macros::EnumDiscriminants`] and [`strum_macros::IntoStaticStr`].
-    fn variant_name(&self) -> &'static str;
-
-    /// A reference to a secondary object providing additional context, if any
-    ///
-    /// This should be [`Some`] if the error happens while evaluating some related object
-    /// (for example: when writing a [`StatefulSet`] owned by your controller object).
-    fn secondary_object(&self) -> Option<ObjectRef<DynamicObject>> {
-        None
-    }
-}
+use super::controller::ReconcilerError;
 
 /// Converts an [`Error`] into a publishable Kubernetes [`Event`]
-fn error_to_event<E: PublishableError>(err: &E) -> Event {
+fn error_to_event<E: ReconcilerError>(err: &E) -> Event {
     // Walk the whole error chain, so that we get all the full reason for the error
     let full_msg = {
         use std::fmt::Write;
@@ -50,7 +30,7 @@ fn error_to_event<E: PublishableError>(err: &E) -> Event {
     };
     Event {
         type_: EventType::Warning,
-        reason: err.variant_name().to_string(),
+        reason: err.category().to_string(),
         note: Some(full_msg),
         action: "Reconcile".to_string(),
         secondary: err.secondary_object().map(|secondary| secondary.into()),
@@ -66,7 +46,7 @@ pub fn publish_controller_error_as_k8s_event<ReconcileErr, QueueErr>(
     controller: &str,
     controller_error: &controller::Error<ReconcileErr, QueueErr>,
 ) where
-    ReconcileErr: PublishableError,
+    ReconcileErr: ReconcilerError,
     QueueErr: Error,
 {
     let (error, obj) = match controller_error {
@@ -103,7 +83,7 @@ mod tests {
     use kube::runtime::reflector::ObjectRef;
     use strum_macros::EnumDiscriminants;
 
-    use super::{error_to_event, PublishableError};
+    use super::{error_to_event, ReconcilerError};
 
     #[derive(Debug, thiserror::Error, EnumDiscriminants)]
     #[strum_discriminants(derive(strum_macros::IntoStaticStr))]
@@ -132,8 +112,8 @@ mod tests {
             }
         }
     }
-    impl PublishableError for ErrorFoo {
-        fn variant_name(&self) -> &'static str {
+    impl ReconcilerError for ErrorFoo {
+        fn category(&self) -> &'static str {
             ErrorFooDiscriminants::from(self).into()
         }
