+++ src/logging/mod.rs
@@ -1,18 +1,8 @@
-use kube::{
-    runtime::{
-        controller::{self, ReconcilerAction},
-        reflector::ObjectRef,
-    },
-    Resource,
-};
 use tracing;
 use tracing_subscriber::EnvFilter;
 
-use crate::{client::Client, logging::k8s_events::publish_controller_error_as_k8s_event};
-
-use self::k8s_events::PublishableError;
-
-pub mod k8s_events;
+pub mod controller;
+mod k8s_events;
 
 /// Initializes `tracing` logging with options from the environment variable
 /// given in the `env` parameter.
@@ -28,45 +18,6 @@ pub fn initialize_logging(env: &str) {
     };
 
     tracing_subscriber::fmt().with_env_filter(filter).init();
-}
-
-/// Reports the controller reconciliation result to all relevant targets
-pub fn report_controller_reconciled<K, ReconcileErr, QueueErr>(
-    client: &Client,
-    controller_name: &str,
-    result: &Result<(ObjectRef<K>, ReconcilerAction), controller::Error<ReconcileErr, QueueErr>>,
-) where
-    K: Resource,
-    ReconcileErr: PublishableError,
-    QueueErr: std::error::Error,
-{
-    match result {
-        Ok((obj, _)) => {
-            tracing::info!(
-                controller.name = controller_name,
-                object = %obj,
-                "Reconciled object"
-            );
-        }
-        Err(err) => report_controller_error(client, controller_name, err),
-    }
-}
-
-/// Reports an error to the operator administrator and, if relevant, the end user
-pub fn report_controller_error<ReconcileErr, QueueErr>(
-    client: &Client,
-    controller_name: &str,
-    error: &controller::Error<ReconcileErr, QueueErr>,
-) where
-    ReconcileErr: PublishableError,
-    QueueErr: std::error::Error,
-{
-    tracing::error!(
-        controller.name = controller_name,
-        error = &*error as &dyn std::error::Error,
-        "Failed to reconcile object",
-    );
-    publish_controller_error_as_k8s_event(client, controller_name, error);
 }
 
 #[cfg(test)]
